We have started to link the CHIMES chemistry solver to the UV radiation from stellar sources when the GALSF_FB_LOCAL_UV_HEATING and CHIMES Config flags are both switched on. For each gas particle, we track the ionising flux (>13.6 eV) and the UV flux in the 6-13.6 eV band (Chimes_fluxPhotIon and Chimes_G0, respectively, in the SphP array), from a number of star age bins set by CHIMES_LOCAL_UV_NBINS in allvars.h (8 by default). We can then specify the UV spectrum of the different age bins and hence the average cross sections of each species, which are given to CHIMES. NOTE: This option has not yet been fully tested.
Additionally, we have added a few options to try to alleviate the 'initial starburst' problem in isolated galaxy simulations, where a strong starburst at the beginning of the simulation tends to destroy the gas disc. These Config flags are listed in Template_Config.sh (they all begin with 'AJR_'). Our preferred solution is to use AJR_STAR_AGE to set the initial stellar age of newly formed stars to 50 Myr in the first 50 Myr, and then to linearly decrease this initial age over the next 100 Myr, so that it is back to normal after 150 Myr. We also use the GALSF_ALT_INIT_STAR option to set all disc stars in the ICs to have a uniform age distribution from 0 to 14 Gyr, while bulge stars are initially set to just 14 Gyr.