Change the location where Makefile.systype gets included
Anything that works now should work the same after this change. The goal is to open up the possibility to do the system-specific settings in the Makefile.systype file. Currently, for example,  if I set CXX = mpic++ in Makefile.systype, it gets overwritten by default settings. This might not be the intended use of Makefile.systype, but it's convenient sometimes.