Added variable temperature floor option.
This new option smoothly varies the temperature floor over some initial
period at the start of the simulation. This can be used with isolated
galaxy simulations to try to smooth out the initial cooling and burst
of star formation, by letting the disc settle down more gently from
an initial temperature of 10^4 K down towards 10 K. This option can be
enabled using the AJR_VARIABLE_TFLOOR Config flag. The parameters
TempFloor_init, TempFloor_final, and TempFloor_time control the initial
and final temperature floors and the initial duration over which the
temperature floor is varied, respectively.

We have also moved the ThermEvolOn flag in CHIMES to the gasVars
structure, which makes it easier to disable cooling for individual
gas particles when OPENMP is included. Additionally, there are minor
updates to CHIMES interpol.c routines.