Added option to reduce SN time-scale at the beginning of the simulation.
This option is intended to alleviate the 'initial starburst' problem in isolated galaxy simulations, where an initial burst of star formation destroys the disc before star formation and stellar feedback become self-regulated. With this option, we reduce the time-scales for type II SNe by some user-specified factor over an initial period t1 (and increase the type II SN rate by the same factor, so that the total number of SNe is conserved). Then, from t1 to t2, we smoothly decrease this factor as 1/t^3 until it reaches unity. This option can be switched on using the AJR_RAPID_SN Config flag. There is also an additional option, AJR_RAPID_SN_LUM, which will increase the luminosity of Type 4 star particles (i.e. those that form within the simulation) by the same factor during this initial period, in addition to reducing the SN time-scales.