Added a Config option AJR_INIT_HALO_METALLICITY.
This option allows us to set two different metallicities for different gas particles, based on their ID. I currently use this to set gas particles in the gaseous halo with a lower metallicity than those in the disc. This option adds two new parameters: InitHaloMetallicity specifies the initial metallicity of this second group of gas particles (in solar units), and HaloID is the lowest particle ID in this group. Then any particles with an ID >= this value will use InitHaloMetallicity, while the others use the original InitMetallicity parameter.