Updates to local radiation treatment in CHIMES.
We have updated the fitting functions for the stellar luminosities of hydrogen-ionising radation, and radiation in the 6-13.6 eV band, which are fit to Starburst99 models. Previously, we were using the Geneva Gv40 Z=0.002 tracks. We now use the Geneva Gv40 Z=0.014 tracks, i.e. at solar metallicity.

We have also added the following options:

CHIMES_Z_DEPENDENT_TAU - This scales the tau_uv and tau_op optical depths around star particles by the gas metallicity. This reflects the fact that these opacities are due to dust grains. This is used when calculating stellar luminosities for both radiation pressure and that go into the chemistry solver for the 6-13.6 eV band.

CHIMES_HI_DEPENDENT_TAU_EUV - This scales the tau_euv optical depth (i.e. for hydrogen-ionising radiation) by the HI fraction around the star particle. Without this option, star particles with ages < 4 Myr end up with zero hydrogen-ionising luminosities, due to the high optical depths as they are still in their dense birth clouds (at least in a M200 = 3e11 Msol isolated disc galaxy that we ran). However, this optical depth is due to absorption by HI, so it should depend on the HI density around the star, and not the total gas density.

CHIMES_OUTPUT_DENS_AROUND_STAR - This writes out the DensAroundStar, MetalDensAroundStar and HIDensAroundStar in the snapshots. This is useful for computing stellar luminosities in post-processing. Requires CHIMES_Z_DEPENDENT_TAU and CHIMES_HI_DEPENDENT_TAU_EUV.

CHIMES_OUTPUT_DELAY_TIME_HII - This writes out the DelayTimeHII in the snapshots. This is useful for identifying gas particles that have been flagged as HII regions in Gizmo. Requires CHIMES_HII_REGIONS or GALSF_FB_HII_HEATING.