Updates to the CHIMES_HII_REGIONS option.
If m_available < m_effective, we now use a random number to stochastically determine whether to ionise the gas particle, as is done in FIRE.

We have also removed the CHIMES_HII_REGION_REFINEMENT option. This did not work, as gas particles went down to time bin 1, even with only moderate refinement. This might be because there was a hard edge between refined and non-refined regions - maybe the refined resolution should be varied gradually over the boundaries. However, we find that this refinement in the HII regions is not actually needed anyway.