Added option to compute initial CHIMES abundances in equilibrium.
This option can be enabled with the CHIMES_INITIALISE_IN_EQM Config flag. For each gas particle, the chemistry solver is integrated at fixed temperature for (1 / nH) Myr (limited to 1 Gyr) ten times. Note that the density gradients, needed for the Sobolev shielding lengths, are already computed before the equilibrium abundances are calculated in init.c. However, the stellar fluxes computed from the gravity tree are not included, because the gravitational forces are not computed until after init(). In other words, we compute equilibrium abundances in the presence of only the extragalactic UV background.