Updated CHIMES implementation to cope with metal enrichment.
When metal enrichment (e.g. from supernovae and stellar winds) is included, we now update the element abundances used within CHIMES, along with the various ion/molecule abundances. When metals are injected into a particle, the abundances of individual metal ions (with respect to hydrogen) are scaled so as to preserve the relative ion fractions of an individual metal.