Added CHIMES_TURB_DIFF_IONS option.
When this option is switched on, it follows the turbulent diffusion of ion and molecule species from the CHIMES module in the same way as the standard turbulent metal diffusion in FIRE. It uses the equivalent of the 'low-order' method from FIRE's turbulent metal diffusion model. This option requires both TURB_DIFF_METALS and TURB_DIFF_METALS_LOWORDER to be defined as well.