Added CHIMES_METAL_DIFFUSION option.
When this option is enabled, the gas-phase abundances of C, N, O, Mg, Si and Fe used within CHIMES for the chemistry and cooling are suppressed using the density-dependent depletion factors from Jenkins, 2009, ApJ, 700, 1299 (with a high-temperature cut off of 10^6 K). We also use the depletion factors of all 16 metals in that paper to compute a density-dependent dust to gas ratio, which we use to scale the dust abundance within CHIMES so that the dust abundance is reduced at low densities, self-consistently with the density-dependent depletion factors.