Added option to record initial masses of star particles.
This option can be switched on using the AJR_RECORD_INITIAL_STELLAR_MASS option. It records the mass of each star particle when it is formed, and writes these out to the snapshots (for Type 4 particles only).

We have also removed the AJR_STAR_AGE option, as this is no longer used.