Implemented custom chimes_exit() routine in the CHIMES module.
CHIMES cannot directly use Gizmo's endrun() routine, as it is defined outside CHIMES. We have therefore modified the CHIMES module such that the User can specify their own custom exit function and then pass this to CHIMES as a function pointer. We have then created a wrapper function around Gizmo's endrun() routine that we can pass to CHIMES in this way.

This means that, if the CHIMES module has to exit, it will effectively be doing this via Gizmo's endrun() routine, rather than just calling exit(EXIT_FAILURE) as we were before.