Fixed bugs when using CHIMES in single precision.
There were many places throughout the code where we had assumed that all of the CHIMES variables were in double precision. However, we now have an option to build CHIMES in single precision. We have therefore made sure that we always cast variables to the ChimesFloat type as needed when passing variables to CHIMES.