Update to the CHIMES_STELLAR_FLUXES.
When adding the stellar fluxes via the gravity tree, the standard FIRE models skip star particles older than 100 Myr. However, in CHIMES we partition the fluxes according to stellar age, which include a bin for ages >100 Myr. We have therefore modified this such that, when CHIMES_STELLAR_FLUXES is switched on, are included. Note that this will also affect the radiation pressure etc.