Fixed a bug in the CHIMES module.
When compiling CHIMES in single precision with an optimisation flag of -O2 or -O3, the integration of the chemistry and cooling would sometimes fail if the particle reached the temperature floor. To fix this, we now use the temp_floor_mode = 1 option, which ends the integration as soon as it reaches the temperature floor, and we ensure that the temperature floor used in CHIMES is no less than 10.1 K, so that it doesn't go below the lowest temperature bin of the rate coefficient tables.