Fixed a bug in the CHIMES module.
When using CHIMES in single precision, the output vector from CVode would sometimes introduce nan's. We now check for this at the beginning of the right-hand-side function before they propagate through into the abundance array itself. If nan's are found, it will return a recoverable error, so that CVode can then adjust and try to correct for this.